Std\Constants + Float literal fix + CAF auto-force generalization
== Std\Constants\Num ==

intMax, intMin, int32Max, int32Min, floatMax, floatMin, floatEpsilon,
floatInfinity, floatNaN. All pure-Yona literals. `intMin` is spelled
`-intMax - 1` because the literal `9223372036854775808` overflows i64
(lexer parses the unsigned magnitude before applying unary minus —
a separate parser quirk not worth fixing just for this one constant).

== Std\Constants\Math ==

pi, tau, e, phi, sqrt2, sqrt3, ln2, ln10, log2e, log10e. IEEE-754
doubles, full precision. Std\Math.pi now re-exports from here.

== Std\Constants\Platform ==

pathSeparator, pathListSeparator, lineSeparator, nullDevice, pageSize,
cacheLineSize, wordSize, cpuCount, pathMax, nameMax, isLittleEndian,
endianness (:little | :big), osName, arch. Platform-sensitive values
call tiny C helpers in src/runtime/platform/os_linux.c (sysconf /
pathconf / uname); everything else is a pure literal.

== FloatExpr: float → double ==

Same bug family as the IntegerExpr overflow fix. `FloatExpr` was
templated on `LiteralExpr<float>` and every parser callsite cast the
lexer's double value to `float` before storing it. Anything past
float32's ~3.4e38 rounded to +inf — `1e308` printed as `inf`. Now
FloatExpr uses double end-to-end. `1.797e308` (Num.floatMax) reads back
as ~1.797e+308 instead of inf.

== CAF auto-force, generalized ==

Std\Constants defines things like `pi = 3.14...` — a 0-arity function
whose body is a literal. When you write `pi > 3.14`, you want the
value, not a function pointer. Previously codegen_identifier returned
the Function* in two cases: (1) when the name was in named_values_
with `val != nullptr` (import path), and (2) when it was only in
compiled_functions_ (second reference after a first force). Now both
cases auto-force if the underlying definition is 0-arity with a
non-Promise return. Test:

  test/codegen/stdlib_constants_math.yona   # pi > 3.14 && pi < 3.15
  test/codegen/stdlib_constants_num.yona    # intMax - int32Max

Also: Std\Stream.naturals now uses `import intMax from
Std\Constants\Num in range 0 intMax`, Std\Math.pi re-exports the
Math.pi CAF — no more hard-coded 9223372036854775807 / 3.14159...
duplicated across the stdlib.

204 test cases, 1205 assertions, all passing.

+++ include/ast.h
@@ -510,12 +510,12 @@ public:
   [[nodiscard]] AstNodeType get_type() const override { return AST_FALSE_LITERAL_EXPR; };
 };
 
-class FloatExpr final : public LiteralExpr<float> {
+class FloatExpr final : public LiteralExpr<double> {
 private:
   void print(std::ostream &os) const override;
 
 public:
-  explicit FloatExpr(SourceContext token, float value);
+  explicit FloatExpr(SourceContext token, double value);
   template<typename ResultType>
   ResultType accept(const AstVisitor<ResultType> &visitor) const {
     return visitor.visit(const_cast<typename std::remove_const<typename std::remove_pointer<decltype(this)>::type>::type*>(this));
